+++ src/main/java/net/sf/jabref/shared/DBMSConnectionProperties.java
@@ -92,6 +92,17 @@ public class DBMSConnectionProperties {
     }
 
     /**
+     * Compares all properties except the password.
+     */
+    public boolean equals(DBMSConnectionProperties properties) {
+        return this.type.equals(properties.getType())
+                && this.host.equalsIgnoreCase(properties.getHost())
+                && this.port == properties.getPort()
+                && this.database.equals(properties.getDatabase())
+                && this.user.equals(properties.getUser());
+    }
+
+    /**
      *  Gets all required data from {@link SharedDatabasePreferences} and sets them if present.
      */
     private void setFromPreferences(SharedDatabasePreferences prefs) {